+++ src/components/statblock/TraitsDisplay.tsx
@@ -9,21 +9,30 @@ function displayTrait({ name, description }: { name: string, description: string
     )
 }
 
-function TraitsDisplay({ traits, hasPaddlesUpgrade }: { traits: string[], hasPaddlesUpgrade: boolean | undefined }) {
+function replaceCore(name: string, hasMagicCore: boolean | undefined) {
+    if (hasMagicCore) {
+        return name.replace("Whale Oil", "Magic Core")
+    }
+    return name
+}
+
+function TraitsDisplay({ traits, hasPaddlesUpgrade, hasAutoPilot, hasMagicCore }: { traits: string[], hasPaddlesUpgrade: boolean | undefined, hasAutoPilot: boolean, hasMagicCore: boolean | undefined }) {
     const hasPaddles = traits.includes('paddling') || hasPaddlesUpgrade
     return (<>
-        {traits.includes('helmsman') ?
+        {traits.includes('helmsman') && !hasAutoPilot ?
             displayTrait(shiptraits['helmsman'])
+            : null}
+        {traits.includes('helmsman') && hasAutoPilot ?
+            displayTrait(shiptraits['nohelmsman'])
             : null}
         {traits.includes('agile') ?
             displayTrait(shiptraits['agile'])
             : null}
         {traits.includes('sails') && traits.includes('engine') && !hasPaddles ?
-            displayTrait(shiptraits['mixedengine'])
+            displayTrait({ name: replaceCore(shiptraits['mixedengine'].name, hasMagicCore), description: shiptraits['mixedengine'].description })
             : null}
         {traits.includes('sails') && traits.includes('engine') && hasPaddles ?
-            displayTrait(shiptraits['mixedenginepaddling'])
-            : null}
+            displayTrait({ name: replaceCore(shiptraits['mixedenginepaddling'].name, hasMagicCore), description: shiptraits['mixedenginepaddling'].description }) : null}
         {traits.includes('sails') && !traits.includes('engine') ?
             displayTrait(shiptraits['sail'])
             : null}