+++ Changelog.md
@@ -1,5 +1,6 @@
 # V0.0.2
 - Round tonnage/engine operation time/fuel capacity to 2 decimals
+- Statblock displays magic core engine trait, automated pilot trait
 
 # V0.0.1
 - Initial version of website